Move HTML test to test_build_html_copyright.py
