--- tests/roots/test-copyright-multiline/index.rst
@@ -1,3 +1,0 @@
-========================
-test-copyright-multiline
-========================
